refactor(tools): extract two named concerns over the 350-LOC ceiling (#782)
Pure moves, zero logic change. Both files were over the project's 350-line
ceiling; #782 names the two seams to pull out.

input-parse.ts (528 -> 462):
  Extract the path-root breadth guard into root-validation.ts:
  isTooBroadRoot + HOME_TARGETS + afkBreadthTargets. Called from exactly
  three sites (cwd, writeRoots, readRoots); the seam was already clean.

bash-restriction-hook.ts (584 -> 557):
  Extract the AFK_HOME spelling concern into afk-home-refs.ts:
  configuredAfkHome + relocatedAfkSensitiveRoots + the relocated forms
  inside allowlistedFileForms (now afkAllowlistFileForms).

Gate: all 180 direct tests (75 bash-restriction, 105 input-parse) pass
unchanged; full suite 13866 pass; tsc --noEmit and build clean.

Closes #782
